refactor: Drawer Panel component extrac
